## Context

There are known issue in turso-cli about high memory consumption for huge dumps loaded to libsql (see https://github.com/tursodatabase/turso-cli/issues/735).

`sqlite-antlr4-parser` is one of the contributors for this issue. There are several problems with current implementation:
1. It supports only `string` input and doesn't support streaming inputs (`io.Reader`). This forces caller to load all query to the memory
2. During the parsing `SplitStatement` implementation uses `antlr.CommonTokenStream` stream which preserve all previously seen tokens in the buffer

While first problem definitely cause inevitable suboptimal behaviour for streaming data - it's pretty hard to change because `antlr.NewIoStream` also loads all data in memory before processing and streaming adoption requires a lot of changes and tweaks on the caller side.

So, we must concentrate on the second factor which is pretty easy to optimize. 

This PR refactor parsing logic and uses simple `bufferedTokenizer` instead of `antlr.CommonTokenStream`. 

`bufferedTokenizer` just keep a buffer for `K` next tokens in order to support look-ahead access requests and this is enough for parser use case.

Some benchmark numbers (`BenchmarkSplitStatement` test):
```
benchstat old new
goos: linux
goarch: amd64
pkg: github.com/libsql/sqlite-antlr4-parser/sqliteparserutils
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
                 │     old     │                 new                 │
                 │   sec/op    │   sec/op     vs base                │
SplitStatement-8   83.71m ± 2%   55.95m ± 2%  -33.16% (p=0.000 n=32)

                 │     old      │                 new                  │
                 │     B/op     │     B/op      vs base                │
SplitStatement-8   71.26Mi ± 0%   31.89Mi ± 0%  -55.25% (p=0.000 n=32)

                 │     old     │                 new                 │
                 │  allocs/op  │  allocs/op   vs base                │
SplitStatement-8   934.0k ± 0%   327.7k ± 0%  -64.91% (p=0.000 n=32)
```

The main benefit however is not reflected on Go benchmarks because this PR drastically improves working set usage. We can see this in the benchmark logs:
```
$> head -n 10 old
?   	github.com/libsql/sqlite-antlr4-parser/sqliteparser	[no test files]
goos: linux
goarch: amd64
pkg: github.com/libsql/sqlite-antlr4-parser/sqliteparserutils
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkSplitStatement-8   	      14	  77382354 ns/op	74722955 B/op	  933986 allocs/op
--- BENCH: BenchmarkSplitStatement-8
    utils_bench_test.go:22: heap in use: 58769408
$> head -n 10 new
?   	github.com/libsql/sqlite-antlr4-parser/sqliteparser	[no test files]
goos: linux
goarch: amd64
pkg: github.com/libsql/sqlite-antlr4-parser/sqliteparserutils
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
BenchmarkSplitStatement-8   	      20	  56963239 ns/op	33443111 B/op	  327731 allocs/op
--- BENCH: BenchmarkSplitStatement-8
    utils_bench_test.go:23: heap in use: 8306688
```

We estimate from this log that working memory set size reduced 7x from `56 MB` to `8 MB`

## Changes

- Implement `bufferedTokenizer`
- Refactor split statements logic
- Expose `StatementIterator` and `CreateStatementIterator` methods (can be helpful in future to not store all statements and process them in streaming fashion)
- Add `BenchmarkSplitStatement` simple benchmark
- Add test for TEMP triggers

## Testing

- Unit tests are passed