@sivukhin 2024-07-13
@penberg - it seems like I don't have access to the `libsql` namespace...